Reorg header json code
